Fix the bug of $digest has already been in progress
Cause: changing scope variable will make watchers be notified. However, the watchers won't be notify if threads that are changing the scope variable is not inside angular cycle. For this, scope.$digest() should be called. But it is safer to wrap the change in $apply(). I found out that this doesn't work either. My solution is wrapping the change in $timeout.