Fix the bugs of the visibility of navigation arrows doesn't work properly when one of the tab content is empty
Previously, we listen for mouse event on ul sub-element and then we reevaluate the visibility of the arrows by checking mouse coordinate in the event handler. This causes problem ul element might be translated out of the screen. It is better that we should listen for mouse event on the parent element of the ul element instead.