Fix rendering bugs on Chrome
Chrome browser has a problem with rendering on animation. When users slide between the tabs, they might experience some part of the tab content is not completely rendered. This is fixed by scheduling a timer to shake the opacity of the carousel at the end of the animation. This forces Chrome to re-render the carousel and its content.